Logging/services/docker_log_tap/requirements.txt
docker
kafka-python